rpmlib-sys: build a C++ binding instead of a C one (fixes #11, #12)
bindgen was having trouble with several critical files when we were
attempting to build a C binding:

https://github.com/iqlusion-io/crates/issues/11
https://github.com/iqlusion-io/crates/issues/12

However, rpmlib's headers are also set up to supply C++ bindings,
and if we rename our header file from `.h` to `.hpp` it signals
to bindgen that we want to treat the project as C++.

If we do, all the previous problems are magically resolved, and
we are able to bind to all of rpmlib as well as librpmbuild and
librpmsign.

We do introduce one new weird minor bindgen issue (see `lib.rs`).
We should probably switch to whitelisting the structs and
functions the `rpmlib` crate actually uses, rather than having
bindgen build a binding to everything.

But good to know it works!