rpmlib-sys: Use cargo features to control binding/linking rpmbuild + rpmsign
It's unlikely anyone will actually use either of these in the near future,
but it's also nice we're successfully generating bindings for them.

Gate both of them on optional cargo features which have to be opted into.
That way we don't waste time running bindgen on them unless people actually
care, but in CI we can ensure we can still produce bindings for them.