iq-cli: GlobalConfig trait + Secret type
This adds a trait for managing global configuration data and loading it
from TOML files.

It uses `macro_rules!` (for implementation expedience) but should
ideally be moved to a proc macro instead.

It also adds a `Secret` type for storing sensitive configuration data
such as passwords, access tokens, or other credentials.