zeroize: Use core::sync::atomic fences
Uses the `fence` and `compile_fence` functions to ensure LLVM reordering
is fenced using the strictest ordering available (`Ordering::SeqCst`).

Additionally this adds a `ZeroizeWithDefault` trait which implements
`Zeroize` using a volatile set of a type's default. This trait is then
used as the bounds for the `IterMut` and `slice` implementations, which
permits slightly more efficient implementations.

Finally, special cases are added for zeroing byte slices and arrays on
nightly using the `volatile_set_memory` intrinsic, which should provide
the best performance for these cases.