Return errors for undersize decode buffers and trailing whitespace
The length calculations we presently perform are dependent on
whitespace-free inputs.

This adds an (unfortunately non-constant time) check that the *last*
character is *NOT* whitespace. Ideally this would be replaced with a
constant time "trim" function that can slice away leading/trailing
whitespace, but for now this gives a meaningful error.

It also adds checks that buffers are correctly sized to receive the
resulting output, where otherwise it would've panicked for an undersize
buffer. This is a potential DoS vector if subtle-encoding was used on
untrusted inputs.