zeroize: Remove scary language about undefined behavior
The documentation previously included a lengthy essay about potential
caveats in what it can guarantee as mixed volatile/non-volatile writes
were previously assumed to be undefined behavior due to language to that
effect in the official documentation for `write_volatile`.

That language has subsequently been removed after discussion about the
safety implications:

https://github.com/rust-lang/rust/pull/60972

Though generally it's not a good idea to mix volatile and non-volatile
writes when using volatile accesses to communicate with e.g. memory
mapped external devices, for the particular usage pattern in this crate,
i.e. writing a zero-value, it can be considered well-defined.