Refactor to use "flat" path representation
Previously `Path` was represented as a slice of byte slices
(i.e. `&[&[u8]]`) and `PathBuf` as a `Vec<Vec<u8>>`.

These sorts of representations of paths preclude ever being able to impl
safe, non-allocating wrapper types, because there is not a safe
reference conversion from nested `Vec`s to nested slices.

This commit switches to using a 1-byte length prefix on each of the
components. This has the advantages of making reference types easier to
work with and also ensures a canonical wire serialization.

It places a max limit of 256-bytes on each path component, but this
is a reasonable upper bound.

The downside is this temporarily relies on an unsafe pointer conversion
to make everything work, however there is some discussion of how to make
this particular kind of reference conversion safe by means of safe
casting from reference types to transparent wrappers around them:

<https://internals.rust-lang.org/t/pre-rfc-patterns-allowing-transparent-wrapper-types/10229>