gaunt/subtle-encoding/tai64: Remove `failure`
These crates weren't getting much value out of `failure`, and the use of
proc macros to generate what is effectively a match statement on enum
variants which returns a string literal adds a lot to compile times.

Furthermore, `failure` is still using the non-1.0 proc macro crates. It
seems there's upstream work to get this fixed[1], but since we're not
really getting value out of the proc macros as-is, there's not much
reason to bother.

[1]: https://github.com/rust-lang-nursery/failure/pull/319