secrecy: Add (optional) concrete `SecretBytes` type
Zeroizing `Bytes` is tricky because it's explicitly designed to provide
an immutable, shared view of bytes.

There's one escape hatch though: `Bytes::try_mut()`, which takes
ownership of `Bytes`, returning a `BytesMut` if it has an exclusive view
of the data, or the original `Bytes` if other readers still have a
shared view of it. That's ok though, because we don't want to zero out
the bytes until the last reader is finished with it.

Using that, and a hack where the `Bytes` are kept in an `Option` and
moved out with `Option::take` in the drop handler, it's possible for the
drop handler to take ownership of the `Bytes`, attempt a conversion to
`BytesMut`, and if it succeeds calling `BytesMut::zeroize()`.