zeroize: Relax `Zeroize` trait bounds for `Vec`
This changes the trait bounds for the `Zeroize` impl on `Vec<Z>` from:

    Z: Clone + Default + Zeroize

to just:

    Z: Zeroize

This is based on my experience trying to switch `curve25519-dalek` from
`clear_on_drop` to `zeroize` in this PR:

https://github.com/dalek-cryptography/curve25519-dalek/pull/289

Notably there were several places I was unable to use `Zeroizing` in
conjunction with `Vec` because it requires `Clone` and `Default` in
order for the current safe impl of zeroizing the excess capacity in
`Vec`s. IMO this is an implementation detail leaking out which places
undue burden on the user and reduces overall utility.

This is another way of saying: this change effectively reverts #180.
I think we can eventually bring this functionality back using an unsafe
impl which writes zeroes to the excess capacity.

In general my opinion is if you're storing secrets in `Vec`, it's best
to avoid mutating that `Vec` and causing reallocations in the first
place. This is the goal of the `SecretVec` type in `secrecy`:

https://docs.rs/secrecy/latest/secrecy/type.SecretVec.html