anomaly: Add generic `Error` type
Provides an `Error` type containing a `Box<Context<Kind>>` which is
useful for propagating errors. It impls `Deref` to the underlying
`Context`, and also `Display` and `std::error::Error`

Unfortunately, having `anomaly` define this type (rather than the
end-user crate) means that the end-user crate can no longer impl `From`
to do automatic error conversions to this type.

An alternative solution would be to have `anomaly` provide a proc macro
which derives the set of traits the `Error` type impls, allowing each
crate to define their own `Error` type, but have them behave in the
same way.

However, for now, this is better than nothing.