secrecy: have `DebugSecret` take a formatter
Adds a `fmt::Formatter` argument to `DebugSecret::debug_secret`, making
it into a full-fledged equivalent of `Debug::fmt`, but without a `&self`
parameter which thereby ensures that there is no way to expose the
underlying secret value.

This makes it possible to change the default impl to use the recently
added `any::type_name` in order to display the *type* of the secret in
the debug message.