zeroize: clear residual space within Option
Always perform a first pass which writes zero to the entirety of an
`Option`s size to ensure any residual data in the Option is cleared.

This is useful if `Some` value is moved out of the option, making a copy
and setting the `Option` to `None`, but leaving data in memory.

Closes #671.