bip32: add Prefix type; MSRV 1.48+
The `Prefix` type is an abstraction over the 32-bit "version" number
decoded from Base58(Check).

It can be computed from and retains the 32-bit "version" number, but
also stores the ASCII representation of the version as well. This allows
it to provide an `AsRef<str>` impl and a human-meaningful `Display`
impl, as well as a better `Debug` which includes both the string and the
integer serializations.

Additionally it provides `const fn` initializers, allowing for `Prefix`
constants, and provides a base set like `Prefix::XPUB`, `Prefix::XPRV`.