bip32: add PublicKey/ExtendedPublicKey and XPub alias
Adds a `PublicKey` trait and uses it as the bound for the associated
`PrivateKey::PublicKey` type.

Adds `ExtendedPublicKey`, the extended key analogue of
`ExtendedPrivateKey`, and support for serializing/deserializing them
to/from `ExtendedKey`, as well as computing it from an
`ExtendedPrivateKey`.

When the `secp256k1` feature is enabled, adds an `XPub` alias for an
`ExtendedPublicKey` whose `K: PublicKey` is a
`k256::ecdsa::VerifyingKey`.

Finally, augments the test vectors in `tests/bip32_vectors.rs` to
include `xpub...` strings, and ensures keys are derived and serialized
correctly.