Bump `k256` to v0.9; `rand_core` to v0.6
We were previously keeping these back to ensure compatibility with
`tendermint-rs` for use with `bip32`, however one of the interesting
features of `bip32` is that it's generic, which means we can still use
it via simple newtype wrappers that provide the current implementation.

This commit bumps these dependencies so we can release with the latest
versions.