signatory: `Algorithm` enum improvements
The `Algorithm` enum was previously missing `EcdsaNistP256` and
`Ed25519` variants, as it wasn't integral to how the keyring worked.

This commit adds those variants, and also uses `Algorithm` as the one
way to parse PKCS#8 AlgorithmIdentifiers, ensuring that in the future
additional algorithms will always need to be added to it.