reject #[zeroize(drop)] on struct/enum fields, enum variants
Reject the #[zeroize(drop)] attribute on struct and enum fields, as users may be missing a zeroizing Drop implementation
they expected to be generated for them.

For consistency, the attribute is now rejected on enum variants as well, which was previously accepted and generated a
Drop impl for the entire enum, but which may not be what users may have expected, as it could be interpreted that such
an attribute would zeroize only the variants containing the attribute on Drop.