Allow converting a `SecretSlice<S>` into `SecretBox<[S; N]>`
This PR allows moving the inner secret of a slice into the inner secret
of an array without changing the actual allocation. If the conversion
fails then the original value is returned.

This prevents a caller from having to clone the secret if they want
to further restrict its type.